+++ ntex-rt/src/lib.rs
@@ -9,6 +9,137 @@ pub use self::arbiter::Arbiter;
 pub use self::builder::{Builder, SystemRunner};
 pub use self::system::System;
 
+#[allow(dead_code)]
+#[cfg(all(feature = "glommio", target_os = "linux"))]
+mod glommio {
+    use std::{future::Future, pin::Pin, task::Context, task::Poll};
+
+    use futures_channel::oneshot::{self, Canceled};
+    use glomm_io::{task, Task};
+    use once_cell::sync::Lazy;
+    use parking_lot::Mutex;
+    use threadpool::ThreadPool;
+
+    /// Runs the provided future, blocking the current thread until the future
+    /// completes.
+    pub fn block_on<F: Future<Output = ()>>(fut: F) {
+        let ex = glomm_io::LocalExecutor::default();
+        ex.run(async move {
+            let _ = fut.await;
+        })
+    }
+
+    /// Spawn a future on the current thread. This does not create a new Arbiter
+    /// or Arbiter address, it is simply a helper for spawning futures on the current
+    /// thread.
+    ///
+    /// # Panics
+    ///
+    /// This function panics if ntex system is not running.
+    #[inline]
+    pub fn spawn<F>(f: F) -> JoinHandle<F::Output>
+    where
+        F: Future + 'static,
+        F::Output: 'static,
+    {
+        JoinHandle {
+            fut: Either::Left(
+                Task::local(async move {
+                    let _ = Task::<()>::later().await;
+                    f.await
+                })
+                .detach(),
+            ),
+        }
+    }
+
+    /// Executes a future on the current thread. This does not create a new Arbiter
+    /// or Arbiter address, it is simply a helper for executing futures on the current
+    /// thread.
+    ///
+    /// # Panics
+    ///
+    /// This function panics if ntex system is not running.
+    #[inline]
+    pub fn spawn_fn<F, R>(f: F) -> JoinHandle<R::Output>
+    where
+        F: FnOnce() -> R + 'static,
+        R: Future + 'static,
+    {
+        spawn(async move { f().await })
+    }
+
+    /// Env variable for default cpu pool size.
+    const ENV_CPU_POOL_VAR: &str = "THREADPOOL";
+
+    static DEFAULT_POOL: Lazy<Mutex<ThreadPool>> = Lazy::new(|| {
+        let num = std::env::var(ENV_CPU_POOL_VAR)
+            .map_err(|_| ())
+            .and_then(|val| {
+                val.parse().map_err(|_| {
+                    log::warn!("Can not parse {} value, using default", ENV_CPU_POOL_VAR,)
+                })
+            })
+            .unwrap_or_else(|_| num_cpus::get() * 5);
+        Mutex::new(
+            threadpool::Builder::new()
+                .thread_name("ntex".to_owned())
+                .num_threads(num)
+                .build(),
+        )
+    });
+
+    thread_local! {
+        static POOL: ThreadPool = {
+            DEFAULT_POOL.lock().clone()
+        };
+    }
+
+    enum Either<T1, T2> {
+        Left(T1),
+        Right(T2),
+    }
+
+    /// Blocking operation completion future. It resolves with results
+    /// of blocking function execution.
+    pub struct JoinHandle<T> {
+        fut: Either<task::JoinHandle<T>, oneshot::Receiver<T>>,
+    }
+
+    impl<T> Future for JoinHandle<T> {
+        type Output = Result<T, Canceled>;
+
+        fn poll(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
+            match self.fut {
+                Either::Left(ref mut f) => match Pin::new(f).poll(cx) {
+                    Poll::Pending => Poll::Pending,
+                    Poll::Ready(res) => Poll::Ready(res.ok_or(Canceled)),
+                },
+                Either::Right(ref mut f) => Pin::new(f).poll(cx),
+            }
+        }
+    }
+
+    pub fn spawn_blocking<F, T>(f: F) -> JoinHandle<T>
+    where
+        F: FnOnce() -> T + Send + 'static,
+        T: Send + 'static,
+    {
+        let (tx, rx) = oneshot::channel();
+        POOL.with(|pool| {
+            pool.execute(move || {
+                if !tx.is_canceled() {
+                    let _ = tx.send(f());
+                }
+            })
+        });
+
+        JoinHandle {
+            fut: Either::Right(rx),
+        }
+    }
+}
+
 #[cfg(feature = "tokio")]
 mod tokio {
     use std::future::Future;
@@ -136,17 +267,37 @@ mod asyncstd {
 #[cfg(feature = "tokio")]
 pub use self::tokio::*;
 
-#[cfg(all(not(feature = "tokio"), feature = "async-std"))]
+#[cfg(all(
+    not(feature = "tokio"),
+    not(feature = "glommio"),
+    feature = "async-std",
+    target_os = "linux"
+))]
 pub use self::asyncstd::*;
+
+#[cfg(all(
+    not(feature = "tokio"),
+    not(feature = "async-std"),
+    feature = "glommio"
+))]
+pub use self::glommio::*;
 
 /// Runs the provided future, blocking the current thread until the future
 /// completes.
-#[cfg(all(not(feature = "tokio"), not(feature = "async-std")))]
+#[cfg(all(
+    not(feature = "tokio"),
+    not(feature = "async-std"),
+    not(feature = "glommio")
+))]
 pub fn block_on<F: std::future::Future<Output = ()>>(_: F) {
     panic!("async runtime is not configured");
 }
 
-#[cfg(all(not(feature = "tokio"), not(feature = "async-std")))]
+#[cfg(all(
+    not(feature = "tokio"),
+    not(feature = "async-std"),
+    not(feature = "glommio")
+))]
 pub fn spawn<F>(_: F) -> std::pin::Pin<Box<dyn std::future::Future<Output = F::Output>>>
 where
     F: std::future::Future + 'static,