+++ ntex/CHANGES.md
@@ -2,6 +2,8 @@
 
 ## [0.5.10] - 2022-01-xx
 
+* rt: Add glommio runtime support
+
 * http: Use Io::take() method for http/1 dispatcher
 
 ## [0.5.9] - 2022-01-12